% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-server.R
\name{apply_schema}
\alias{apply_schema}
\title{Apply a schema to data}
\usage{
apply_schema(data, schema, keys_ignore = NULL)
}
\arguments{
\item{data}{A tibble::tibble contianing values.}

\item{schema}{A named list of character vectors. Each vector is matched to a column in 'data' and transforms it into a factor.}

\item{keys_ignore}{A character vector or NULL, these columns are not transformed.}
}
\value{
A tibble::tibble.
}
\description{
Apply a schema to data
}
