% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-server.R
\name{has_length}
\alias{has_length}
\title{Function to determine if an R object has length}
\usage{
has_length(x)
}
\arguments{
\item{x}{an R object. For replacement, a vector or factor.}
}
\value{
a non-negative integer or double (which will be rounded down).
}
\description{
Function to determine if an R object has length
}
